# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BitGate API - blockchain-verified API marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketplaceApi {
  id: number;
  name: string;
  description: string;
  price: number;
  asset: string;
  endpoint: string;
  usageCount: number;
  earnings: number;
  category: string;
}

export interface VerifyPaymentRequest {
  txid: string;
  userAddress: string;
  apiId: number;
}

export interface VerifyPaymentResponse {
  success: boolean;
  apiKey?: string;
  message?: string;
}

export interface PriceFeed {
  BTC: number;
  STX: number;
  ETH: number;
  timestamp: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetApiKeyParams = {
  address: string;
};