# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BitGate API - blockchain-verified API marketplace
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: marketplace
    description: API marketplace listings
  - name: payment
    description: Payment verification and API key issuance
  - name: protected
    description: Paywalled API endpoints

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /apis:
    get:
      operationId: listApis
      tags: [marketplace]
      summary: List available APIs in the marketplace
      responses:
        "200":
          description: List of marketplace APIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceApi"

  /payment/verify:
    post:
      operationId: verifyPayment
      tags: [payment]
      summary: Verify an on-chain payment and issue an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified and API key issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPaymentResponse"
        "400":
          description: Payment not confirmed or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment/key:
    get:
      operationId: getApiKey
      tags: [payment]
      summary: Get existing API key for an address
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API key found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPaymentResponse"
        "404":
          description: No key found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /price-feed:
    get:
      operationId: getPriceFeed
      tags: [protected]
      summary: Get live crypto price feed (requires API key)
      parameters:
        - name: x-api-key
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Price data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceFeed"
        "401":
          description: API key required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    MarketplaceApi:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        price:
          type: number
        asset:
          type: string
        endpoint:
          type: string
        usageCount:
          type: number
        earnings:
          type: number
        category:
          type: string
      required: [id, name, description, price, asset, endpoint, usageCount, earnings, category]

    VerifyPaymentRequest:
      type: object
      properties:
        txid:
          type: string
        userAddress:
          type: string
        apiId:
          type: number
      required: [txid, userAddress, apiId]

    VerifyPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        apiKey:
          type: string
        message:
          type: string
      required: [success]

    PriceFeed:
      type: object
      properties:
        BTC:
          type: number
        STX:
          type: number
        ETH:
          type: number
        timestamp:
          type: number
      required: [BTC, STX, ETH, timestamp]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]