# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BitGate API - blockchain-verified API marketplace
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List available APIs in the marketplace
 */
export const ListApisResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  asset: zod.string(),
  endpoint: zod.string(),
  usageCount: zod.number(),
  earnings: zod.number(),
  category: zod.string(),
});
export const ListApisResponse = zod.array(ListApisResponseItem);

/**
 * @summary Verify an on-chain payment and issue an API key
 */
export const VerifyPaymentBody = zod.object({
  txid: zod.string(),
  userAddress: zod.string(),
  apiId: zod.number(),
});

export const VerifyPaymentResponse = zod.object({
  success: zod.boolean(),
  apiKey: zod.string().optional(),
  message: zod.string().optional(),
});

/**
 * @summary Get existing API key for an address
 */
export const GetApiKeyQueryParams = zod.object({
  address: zod.coerce.string(),
});

export const GetApiKeyResponse = zod.object({
  success: zod.boolean(),
  apiKey: zod.string().optional(),
  message: zod.string().optional(),
});

/**
 * @summary Get live crypto price feed (requires API key)
 */
export const GetPriceFeedHeader = zod.object({
  "x-api-key": zod.string().optional(),
});

export const GetPriceFeedResponse = zod.object({
  BTC: zod.number(),
  STX: zod.number(),
  ETH: zod.number(),
  timestamp: zod.number(),
});