# lib/api-zod/src/generated/types/marketplaceApi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BitGate API - blockchain-verified API marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceApi {
  id: number;
  name: string;
  description: string;
  price: number;
  asset: string;
  endpoint: string;
  usageCount: number;
  earnings: number;
  category: string;
}