# lib/api-zod/src/generated/types/priceFeed.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BitGate API - blockchain-verified API marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface PriceFeed {
  BTC: number;
  STX: number;
  ETH: number;
  timestamp: number;
}